fix(builtin): attempt to fix race with package.json and also gather more information
We're randomly seeing malformed package.json files during a clean build.
The package.json files in the repo npm install and npm ci fine. However,
rarely during a clean build there will be a SyntaxError about reading a
package.json in Bazel land. There seems to be a syntax error in exactly
the same place in the same file.

This fix adds a retry with a delay to reading package.json files in case
there's any kind of concurrency error. I doubt there is, but who knows.

This also adds extra logging, so we can get more information about what
is going on.