fix(builtin): attempt to fix potential race with npm files and the OS file buffer
We're randomly seeing malformed package.json files during a clean build.
The package.json files in the repo npm install and npm ci fine. However,
rarely during a clean build there will be a SyntaxError about reading a
package.json in Bazel land. The hypothesis is there's so many files
being written by npm ci/install and this code is quickly enough after
npm exits, that there may be files not be fully written to disk.

This is an attempt to make sure any package.json files are flushed to
the disk before reading them.

Not totally sure this is the root cause of the issue, but I figure it's
worth a try.