fix(builtin): stop using removeNPMAbsolutePaths as it is no longer needed with modern versions of npm
This may also help with a malformed package.json file mysteriously and
infrequently being produced.

Node removed this tool in late 2021 saying it was no longer needed
https://github.com/nodejs/node/commit/6525226ff7 and the repo for the
tool hasn't been updated since 2022. There is a closed issue on that
repo suggesting it is no longer needed for modern versions of npm.

Considering that, I think we should be good to remove it and not
introduce any determinism problems.