fix(runfiles): fixed some issues with temporary directory creation
The directories we were creating weren't random, so it was possible tha
two Node actions could share the same directory and one could delete it
before the other.

Also, when catching process events like `uncaughtException` and
signal events, we need explicitly stop the process (and make sure we
return the correct code). In the case of `uncaughtException`, we also
need to handle the error thrown in the same way Node.js would've. Doing
clean up work seems to be an very difficult problem in Node (hence the
existence of `https://github.com/nodejs/node/issues/20804`), so I've
decided to just catch `exit` for the time being. Sure, we may leak
temporary directories if a certain signal is sent, but I think the risk
of not handling signals correctly outweighs the risk of leaking
temporary directories.