fix(runfiles): write Node to the temporary directory
We have a Node program in our private monorepo that looks at the
zeroeth argument (the path to the Node binary) and uses that to spawn a
subprocess. The problem with that is that prior to this change, the
Node.js binary was located in a temporary directory relative to the
output directory of the `nodejs_binary`. This commit changes it to
create a temporary directory in `/tmp`. I thought it was doing this
before, but didn't realize that `fs.mkdtemp` doesn't use the
temporary directory by default.